Update coefficients for PWR/MG
