+++ src/lib/mechanics/damage.ts
@@ -32,7 +32,7 @@ export function calculateDamage(
 
   const isSchismActive = hasAura(state, "Schism");
   const isScovActive = hasAura(state, "Shadow Covenant");
-  const schismMultiplier = isSchismActive ? 1.25 : 1;
+  const schismMultiplier = isSchismActive ? 1.15 : 1;
   const scovMultiplier = isScovActive ? 1.25 : 1;
   const sinsMultiplier = calculateSinsDamageBonus(state);
   const { player } = state;
